requirements.txt
numpy>=1.24
pyyaml>=6.0
pyglet==2.1.8
PyOpenGL==3.1.6
pytest>=7.0
reportlab>=3.6
matplotlib>=3.8
